==> notebooks/03-train-model.ipynb <==
# Model Training Notebook

Clean and extract features from raw data

# Steps

1. Split the data into training and test data set
1. Clean the data (transform null values)
1. Scale necessary attributes (normalization, standardization)
1. Save transformed data for model training


# Import packages

# model
from sklearn import svm
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression

# serializing, compressing, and loading the models
import joblib
from tabulate import tabulate
# performance
from sklearn.model_selection import cross_val_score
from sklearn.metrics import f1_score

# displaying plots
import numpy as np
import sys
sys.path.append("../lib")

from getConfig import *
config = getConfig("../")
config.cleanup(config.trained_path)


# Model Selection

While there are several classifiers available, we show how to train the following classifiers, compare and select one.

[Classification Model Comparison](https://scikit-learn.org/stable/auto_examples/classification/plot_classifier_comparison.html?highlight=svm%20svc)

1. [Support Vector Classifier](https://scikit-learn.org/stable/modules/generated/sklearn.svm.SVC.html#sklearn.svm.SVC)
1. [Random Forest Classifier](https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestClassifier.html#sklearn.ensemble.RandomForestClassifier)
1. [Logistic Regression](https://scikit-learn.org/stable/modules/linear_model.html#logistic-regression)

If you are interested in a more extensive collection, i.e, training other kinds of models and comparing them to pick the best one, please refer

https://scikit-learn.org/stable/tutorial/machine_learning_map/index.html

### Model Evaluation
https://scikit-learn.org/stable/modules/model_evaluation.html

In this notebook, we use the following APIs to evaluate the quality of a model’s predictions. For more detailed evaluation and other metrics, please refer to the link above

1. Estimator score method (Score)
2. Scoring parameter (Cross validation)
3. Metric functions (F1 score from Classification metrics)

Best practice to save every model you experiment with so you can come back easily to any model.
Save both the hyperparameters and trained parameter, as well as the cross-validation scores and predictions.
This will allow you to easily compare scores across model types. Use Pickle or joblib libraries.

## Load training and test data


with open(config.traintest_path + "X_train_prepared.csv") as file_name:
    X_train_prepared = np.loadtxt(file_name, delimiter=",")

with open(config.traintest_path + "X_train_prepared_m.csv") as file_name:
    X_train_prepared_m = np.loadtxt(file_name, delimiter=",")
    
with open(config.traintest_path + "y_train.csv") as file_name:
    y_train = np.loadtxt(file_name, delimiter=",")
    
with open(config.traintest_path + "y_test.csv") as file_name:
    y_test = np.loadtxt(file_name, delimiter=",")




## Initialize results and define method to evaluate models


# Initialize rows and columns of table to store results

results = [["Support Vector Classifier"], 
          ["Random Forest Classifier"], 
          ["Logistic Regression"]]
  
#define header names
col_names = ["Estimator score method", "Scoring parameter", "Metric functions"]


def eval_model(model):

    # ESTIMATOR SCORE METHOD

    accuracy = model.score(X_train_prepared,y_train)

    # SCORING PARAMETER
    cross_validation = cross_val_score(model, X_train_prepared, y_train, cv=3, scoring='recall_macro')

    # METRIC FUNCTIONS

    y_pred = model.predict(X_train_prepared)
    f1score = f1_score(y_train, y_pred)

    print ('Accuracy:' + str(accuracy) +'\nCross validation score: ' + str(cross_validation) + '\nf1 score: ' + str (f1score))

    var = [accuracy, cross_validation, f1score]
    return var

## Support Vector Classifier (SVC)
https://scikit-learn.org/stable/modules/generated/sklearn.svm.SVC.html#sklearn.svm.SVC

### Model Training

# select the SVM
clf = svm.SVC()

# fit the model to the data
CLF = clf.fit(X_train_prepared, y_train)


### Model Evaluation

result0 = eval_model(CLF)
for item in result0:
    results[0].append(item)

### Store Model

# save and compress the model

joblib.dump(CLF, config.trained_path + "svc_model.pkl", compress=('bz2', 3))

## Random Forest Classifier

### Model Training

# select the estimator
CLF = RandomForestClassifier()

# fit the model to the data
CLF.fit(X_train_prepared, y_train)

### Model Evaluation

result1 = eval_model(CLF)
for item in result1:
    results[1].append(item)

### Store Model

# save the model
joblib.dump(CLF, config.trained_path + "rfc_model.pkl", compress=('bz2', 3))


## Logistic Regression
https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html?highlight=logisticregression#sklearn.linear_model.LogisticRegression)

### Model Training

# select the classifier
CLF = LogisticRegression()

# fit the model to the data
CLF.fit(X_train_prepared, y_train)

### Model Evaluation

result2 = eval_model(CLF)
for item in result2:
    results[2].append(item)

### Store Model

# save the model
joblib.dump(CLF, config.trained_path + "log_model.pkl", compress=('bz2', 3))

# Model Selection

#display table
print(tabulate(results, headers=col_names))


fp = open(config.trained_path + "model_comparison.txt", "w")
fp.write(tabulate(results, headers=col_names))
fp.close()


# Launch, Monitor, and Maintain

1. Use joblib to save the train model inclulding full pre-processing and prediction pipeline
1. Load the trained model to production
1. Call the predict() method to make predictions

## Serving
1. Load the model in a web app that will call the predict() method
1. Wrap the model in a dedicated web service that a web app queries with REST API
    1. makes it easy to upgrade without interrupting the primary web app
    1. makes it easy to scale web services and load balance the requests from the web app across the web services
    1. enables the web app to use any language, not just Python

## Monitor
1. Write monitor code the check live performance at regular intervals and trigger alerts when it drops
    1. Could be steep drop if an infrastructure components stops
    1. or, a gentle decay as the world changes resulting in model rot
